+++ code/inc/graphics/core/IGraphics.hpp
@@ -5,8 +5,7 @@
 class IGraphics
 {
 public:
-
 	// [[ INTERFACE METHODS ]]
 	virtual void Init(ISurface* pSurface) = 0;
-	virtual void Draw(char c, uint16 x, uint16 y) = 0;
+	virtual void Draw(char c, int32 x, int32 y) = 0;
 };
