+++ code/src/graphics/CharGraphics.cpp
@@ -25,7 +25,7 @@ void CharGraphics::Init(ISurface* pSurface)
 	}
 }
 
-void CharGraphics::Draw(char c, uint16 x, uint16 y)
+void CharGraphics::Draw(char c, int32 x, int32 y)
 {
 	// Flip the indices so we can access a row at a time more easily when outputting to the console.
 	m_pSurface->GetBackBuffer().ppData[y][x] = c;
